requirements.txt
uvicorn
fastapi
pydantic
python-dateutil
sqlalchemy
passlib
jwt

black